# v0/openapi/aihub_openapi.yaml
openapi: 3.1.0
info:
  title: ieops-aihub
  version: 0.4.0
  description: |
    polyforge-v3 backend. v3-design.md §5 (schema) + §10.1 (MCP tools)。
    所有 mutating endpoint 强制带 AttemptCredential 三联 (attempt_id, claim_epoch, session_secret)。
servers:
  - url: http://10.146.0.16
    description: GMI 内网默认部署 (design §6.1 推荐 HTTPS for 公网, warn-only on http://非localhost)
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: aihub_sk_*
  responses:
    Unauthorized:
      description: 401 — invalid api_key / session_secret / lease
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
    Forbidden:
      description: 403 — actor mismatch / not authorized for project
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
    NotFound:
      description: "404 — resource not found"
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
    BadRequest:
      description: 400 — payload schema 校验失败
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
    Conflict:
      description: 409 — epoch mismatch / lease expired / lock hard_block / busy_not_takeover_eligible
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
    PayloadTooLarge:
      description: 413 — event payload > 64 KiB (design §20 M5)
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
  schemas:
    AttemptCredential:
      type: object
      required: [attempt_id, claim_epoch, session_secret]
      properties:
        attempt_id: {type: string, pattern: "^ra_[a-z0-9]+$"}
        claim_epoch: {type: integer, minimum: 1}
        session_secret:
          type: string
          pattern: "^[0-9a-f]{64}$"
          minLength: 64
          maxLength: 64
          description: "raw 32-byte hex (64 chars); server sha256 比对存储的 session_secret_hash (§6.1)"
    ErrorEnvelope:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - UNAUTHORIZED
            - FORBIDDEN
            - CONFLICT_EPOCH_MISMATCH
            - CONFLICT_LEASE_EXPIRED
            - CONFLICT_LEASE_TAKEN_OVER
            - CONFLICT_HARD_BLOCK
            - CONFLICT_BUSY_NOT_TAKEOVER_ELIGIBLE
            - CONFLICT_DUPLICATE_REQUEST
            - BAD_REQUEST
            - NOT_FOUND
            - PAYLOAD_TOO_LARGE
            - INTERNAL_ERROR
            - SERVICE_UNAVAILABLE
            - UNKNOWN_REMOTE_ERROR
          description: "13 known ErrorCode (§6.1) + UNKNOWN_REMOTE_ERROR forward-compat placeholder; SoT polyforge_v3.aihub.errors.ErrorCode"
        message: {type: string}
        details: {type: object, additionalProperties: true}
...
    Memory:
      type: object
      required: [id, project, author_user_id, visibility, type, metadata, created_at]
      properties:
        id: {type: string}
        project: {type: string}
        author_user_id: {type: string}
        work_item_id: {type: [string, "null"]}
        visibility: {type: string, enum: [private, project, team, admin]}
        type: {type: string}
        content: {type: [string, "null"]}
        metadata: {type: object}
        embedding: {type: [array, "null"], items: {type: number}}
        redacted_at: {type: [string, "null"], format: date-time}
        redaction_reason: {type: [string, "null"]}
        expires_at: {type: [string, "null"], format: date-time}
        created_at: {type: string, format: date-time}
    AgentEvent:
      type: object
      required: [id, work_item_id, actor_user_id, event_type, payload, created_at]
      properties:
        id: {type: string}
        work_item_id: {type: string}
        run_attempt_id: {type: [string, "null"]}
        actor_user_id: {type: string}
        api_key_id: {type: [string, "null"]}
        event_type: {type: string}
        payload: {type: object}
        pinned: {type: boolean}
        created_at: {type: string, format: date-time}
...
    EmitEventRequest:
      allOf:
        - $ref: '#/components/schemas/AttemptCredential'
        - type: object
          required: [work_item_id, event_type, payload]
          properties:
            work_item_id: {type: string}
            event_type: {type: string}
            payload: {type: object}
            pinned: {type: boolean}
...
    CreateMemoryRequest:
      type: object
      required: [project, type, content, visibility]
      properties:
        project: {type: string}
        type: {type: string}
        content: {type: string}
        visibility: {type: string, enum: [private, project, team, admin]}
        work_item_id: {type: string}
        metadata: {type: object}
        expires_at: {type: string, format: date-time}
    UpdateMemoryRequest:
      type: object
      required: [patch_payload]
      properties:
        patch_payload: {type: object}
    RedactMemoryRequest:
      type: object
      description: "Per §6.1 memory carve-out — 不带 AttemptCredential (memory 是 per-user 不是 per-attempt)"
      required: [reason]
      properties:
        reason: {type: string, description: "redaction 原因, 落 memories.redaction_reason"}
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok: {type: boolean}
    EmitEventResponse:
      type: object
      required: [event_id]
      properties:
        event_id: {type: string}
...
    EventListResponse:
      type: object
      required: [items, next_cursor]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/AgentEvent'}}
        next_cursor: {type: [string, "null"]}
    MemoryListResponse:
      type: object
      required: [items, next_cursor]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/Memory'}}
        next_cursor: {type: [string, "null"]}
paths:
...
  /v1/events:
    post:
      operationId: emit_event
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/EmitEventRequest'}}}
      responses:
        '201':
          description: event recorded
          content: {application/json: {schema: {$ref: '#/components/schemas/EmitEventResponse'}}}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '409': {$ref: '#/components/responses/Conflict'}
        '413': {$ref: '#/components/responses/PayloadTooLarge'}
    get:
      operationId: list_events
      parameters:
        - {name: work_item_id, in: query, schema: {type: string}}
        - {name: types, in: query, schema: {type: string, description: "comma-separated event_type list"}}
        - {name: since, in: query, schema: {type: string, format: date-time}}
        - {name: cursor, in: query, schema: {type: string}}
        - {name: limit, in: query, schema: {type: integer, default: 50, maximum: 200}}
      responses:
        '200':
          description: events list
          content: {application/json: {schema: {$ref: '#/components/schemas/EventListResponse'}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
...
  /v1/memories:
    post:
      operationId: create_memory
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/CreateMemoryRequest'}}}
      responses:
        '201':
          description: created
          content: {application/json: {schema: {$ref: '#/components/schemas/Memory'}}}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
    get:
      operationId: list_memories
      parameters:
        - {name: project, in: query, schema: {type: string}}
        - {name: query, in: query, schema: {type: string}}
        - {name: type, in: query, schema: {type: string}}
        - {name: visibility, in: query, schema: {type: string}}
        - {name: work_item_id, in: query, schema: {type: string}}
        - {name: top_k, in: query, schema: {type: integer}}
        - {name: cursor, in: query, schema: {type: string}}
        - {name: limit, in: query, schema: {type: integer, default: 50, maximum: 200}}
      responses:
        '200':
          description: list
          content: {application/json: {schema: {$ref: '#/components/schemas/MemoryListResponse'}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /v1/memories/{memory_id}:
    parameters:
      - {name: memory_id, in: path, required: true, schema: {type: string}}
    patch:
      operationId: update_memory
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/UpdateMemoryRequest'}}}
      responses:
        '200':
          description: updated
          content: {application/json: {schema: {$ref: '#/components/schemas/Memory'}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /v1/memories/{memory_id}/redact:
    parameters:
      - {name: memory_id, in: path, required: true, schema: {type: string}}
    post:
      operationId: redact_memory
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/RedactMemoryRequest'}}}
      responses:
        '200':
          description: redacted
          content: {application/json: {schema: {$ref: '#/components/schemas/OkResponse'}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
...